3. Интерфейс Input [#181092]

+++ src/main/java/ru/job4j/poly/Transport.java
@@ -5,6 +5,6 @@ public interface Transport {
 
     void passengers(int number);
 
-    double refuel(double volume, int price);
+    double refuel(double volume, double price);
 
 }
